+++ routes/users.js
@@ -1,10 +1,12 @@
 const express = require("express");
-const { check, validationResult } = require("express-validator");
+const { validationResult } = require("express-validator");
 const bcrypt = require("bcryptjs");
 
 const db = require("../db/models");
 const { csrfProtection, asyncHandler } = require("../utils");
 const { loginUser, logoutUser } = require("../auth");
+const userValidators = require("../validators/user-validators");
+const loginValidators = require("../validators/login-validators");
 
 const router = express.Router();
 
@@ -16,60 +18,6 @@ router.get("/new", csrfProtection, (req, res) => {
     csrfToken: req.csrfToken(),
   });
 });
-
-const userValidators = [
-  check("username")
-    .exists({ checkFalsy: true })
-    .withMessage("Please enter Username")
-    .isLength({ max: 30 })
-    .withMessage("Username must not be more than 30 characters long"),
-  check("firstName")
-    .exists({ checkFalsy: true })
-    .withMessage("Please enter First Name")
-    .isLength({ max: 50 })
-    .withMessage("First Name must not be more than 50 characters long"),
-  check("lastName")
-    .exists({ checkFalsy: true })
-    .withMessage("Please enter Last Name")
-    .isLength({ max: 50 })
-    .withMessage("Last Name must not be more than 50 characters long"),
-  check("email")
-    .exists({ checkFalsy: true })
-    .withMessage("Please enter Email Address")
-    .isLength({ max: 255 })
-    .withMessage("Email Address must not be more than 255 characters long")
-    .isEmail()
-    .withMessage("Email Address is not a valid email")
-    .custom((value) => {
-      return db.User.findOne({ where: { email: value } }).then((user) => {
-        if (user) {
-          return Promise.reject(
-            "The provided Email Address is already in use by another account"
-          );
-        }
-      });
-    }),
-  check("password")
-    .exists({ checkFalsy: true })
-    .withMessage("Please enter Password")
-    .isLength({ max: 50 })
-    .withMessage("Password must not be more than 50 characters long")
-    .matches(/^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#$%^&*])/, "g")
-    .withMessage(
-      'Password must contain at least 1 lowercase letter, uppercase letter, number, and special character (i.e. "!@#$%^&*")'
-    ),
-  check("confirmPassword")
-    .exists({ checkFalsy: true })
-    .withMessage("Please confirm Password")
-    .isLength({ max: 50 })
-    .withMessage("Confirm Password must not be more than 50 characters long")
-    .custom((value, { req }) => {
-      if (value !== req.body.password) {
-        throw new Error("Confirm Password does not match Password");
-      }
-      return true;
-    }),
-];
 
 router.post(
   "/new",
@@ -111,15 +59,6 @@ router.get("/sign-in", csrfProtection, (req, res) => {
     csrfToken: req.csrfToken(),
   });
 });
-
-const loginValidators = [
-  check("email")
-    .exists({ checkFalsy: true })
-    .withMessage("Please provide a value for Email"),
-  check("password")
-    .exists({ checkFalsy: true })
-    .withMessage("Please provide a value for Password"),
-];
 
 router.post(
   "/sign-in",
