+++ validators/login-validators.js
@@ -1,0 +1,12 @@
+const { check } = require("express-validator");
+
+const loginValidators = [
+  check("email")
+    .exists({ checkFalsy: true })
+    .withMessage("Please provide a value for Email"),
+  check("password")
+    .exists({ checkFalsy: true })
+    .withMessage("Please provide a value for Password"),
+];
+
+module.exports = loginValidators;
